requirements.txt
python-telegram-bot==20.7
supabase==2.0.2
openai==1.3.8
requests==2.31.0
python-dotenv==1.0.0
validators==0.22.0
pytz==2023.3
